feat: add `Fetch` component
+ Create generic `FetchFactory` component
+ Share `FetchFactory` between new `Fetch` and `MQL`
