+++ packages/app/src/components/fetch.js
@@ -1,0 +1,19 @@
+import ky from '@microlink/mql/src/ky'
+
+import { FetchFactory } from './fetch-factory'
+
+const pending = new Map()
+
+const promiseHandler = promise => promise.json()
+
+export const Fetch = ({ children, url, options }) => (
+  <FetchFactory
+    cache={pending}
+    func={ky}
+    url={url}
+    options={options}
+    promiseHandler={promiseHandler}
+  >
+    {children}
+  </FetchFactory>
+)
