+++ packages/app/src/components/index.js
@@ -10,7 +10,6 @@ export { InternalLink } from './internal-link'
 export { HorizontalDragBar, VerticalDragBar } from './drag-bars'
 export { JSONViewer } from './json-viewer'
 export { Label } from './label'
-export { MQL } from './mql'
 export { Script } from './script'
 export { LiveEditor, LiveError, LivePreview, LiveProvider } from './live-editor'
 export { Overlay } from './overlay'
