+++ packages/app/src/components/mql.js
@@ -1,31 +1,16 @@
-import { useState, useEffect } from 'react'
-
-import { useMountedRef } from '@/hooks'
 import mql from '@microlink/mql'
+
+import { FetchFactory } from './fetch-factory'
 
 const pending = new Map()
 
-export const MQL = ({ children, url, options } = {}) => {
-  const [result, setResult] = useState(null)
-  const isMounted = useMountedRef()
-
-  useEffect(() => {
-    async function fetchData () {
-      const id = JSON.stringify({ url, ...options })
-
-      let promise = pending.get(id)
-
-      if (!promise) {
-        promise = mql(url, options)
-        pending.set(id, promise)
-      }
-
-      const result = await promise
-      if (isMounted.current) setResult(result)
-    }
-
-    fetchData()
-  }, [isMounted, url, options])
-
-  return children(result)
-}
+export const MQL = ({ children, url, options }) => (
+  <FetchFactory
+    cache={pending}
+    func={mql}
+    url={url}
+    options={options}
+  >
+    {children}
+  </FetchFactory>
+)
