+++ packages/app/src/components/presets/scope.js
@@ -2,6 +2,7 @@ import styled from 'styled-components'
 import * as polished from 'polished'
 import Microlink from '@microlink/react'
 
+import { Fetch } from '../fetch'
 import { Link } from '../link'
 import { Lighthouse } from '../lighthouse'
 import { MQL } from '../mql'
@@ -10,4 +11,4 @@ import { Script } from '../script'
 // eslint-disable-next-line
 export * from 'theme-ui'
 
-export { Microlink, Script, MQL, polished, Link, Lighthouse, styled }
+export { Fetch, Link, Lighthouse, Microlink, MQL, Script, polished, styled }
